requirements.txt
Django~=4.2.7
django-extensions==3.2.1
django-crispy-forms~=2.1